Add a short readme
Until now, this was a bare repo with just the raw markdown content for
the knurling sessions and no instructions on how to access the rendered
version.

This change adds a readme that adds at least a little bit of context for
people that are not contributors to this repo. It explains how to access
the rendered version, the credentials and the current work in progress
state of the book.